# docs/relay/relay_swagger.yaml
basePath: /
definitions:
...
  github_com_kwilabs_straw-proxy-server_internal_server_dto.CacheStatsResponse:
    description: Response containing Redis cache statistics
    properties:
      info:
        description: Info contains Redis server information
        type: string
    type: object
  github_com_kwilabs_straw-proxy-server_internal_server_dto.ClearCacheResponse:
    description: Response after clearing cache keys
    properties:
      deleted:
        description: Deleted is the number of keys removed
        type: integer
      message:
        description: Message describes what was done
        type: string
      pattern:
        description: Pattern is the key pattern that was matched
        type: string
    type: object
...
  github_com_kwilabs_straw-proxy-server_internal_server_dto.CreateFingerprintRequest:
    description: Request body for creating or updating a fingerprint preset
    properties:
      config:
        additionalProperties: true
        description: Config is the detailed configuration of the fingerprint
        type: object
      id:
        description: ID is the unique identifier for this preset (required)
        type: string
      name:
        description: Name is a human-readable name for the preset (required)
        type: string
    required:
    - id
    - name
    type: object
...
  github_com_kwilabs_straw-proxy-server_internal_server_dto.EndpointHealthResponse:
    description: Health information for a registered endpoint
    properties:
      active_tasks:
        description: ActiveTasks is the number of in-flight requests
        type: integer
      endpoint_id:
        description: EndpointID is the unique identifier
        type: string
      last_seen:
        description: LastSeen is when the endpoint last reported in
        type: string
      state:
        description: State is the current health state
        type: string
      tags:
        description: Tags are the endpoint's capability tags
        items:
          type: string
        type: array
      version:
        description: Version is the endpoint software version
        type: string
    type: object
...
  github_com_kwilabs_straw-proxy-server_internal_server_dto.FingerprintResponse:
    description: Fingerprint preset data returned by the API
    properties:
      config:
        additionalProperties: true
        description: Config is the detailed configuration of the fingerprint
        type: object
      created_at:
        description: CreatedAt is when the preset was created
        type: string
      id:
        description: ID is the unique identifier for this preset
        type: string
      name:
        description: Name is a human-readable name for the preset
        type: string
      updated_at:
        description: UpdatedAt is when the preset was last modified
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: Straw Proxy Support
    url: https://github.com/kwilabs/straw-proxy-server
  description: |-
    HTTP relay proxy with fingerprinting and routing capabilities.
    This API allows clients to proxy HTTP requests through distributed endpoints.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Straw Proxy Relay Server API
  version: "1.0"
paths:
...
  /cache/clear:
    post:
      description: Clears Redis cache keys matching the specified pattern
      parameters:
      - description: 'Key pattern to match (default: *)'
        in: query
        name: pattern
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deletion result
          schema:
            $ref: '#/definitions/github_com_kwilabs_straw-proxy-server_internal_server_dto.ClearCacheResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKeyAuth: []
      summary: Clear Cache
      tags:
      - cache
  /cache/stats:
    get:
      description: Returns Redis server information and memory statistics
      produces:
      - application/json
      responses:
        "200":
          description: Redis info
          schema:
            $ref: '#/definitions/github_com_kwilabs_straw-proxy-server_internal_server_dto.CacheStatsResponse'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKeyAuth: []
      summary: Get Cache Stats
      tags:
      - cache
  /endpoints:
    get:
      description: Returns all registered proxy endpoints with their health status
      produces:
      - application/json
      responses:
        "200":
          description: List of endpoints
          schema:
            items:
              $ref: '#/definitions/github_com_kwilabs_straw-proxy-server_internal_server_dto.EndpointHealthResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKeyAuth: []
      summary: List Endpoints
      tags:
      - endpoints
  /endpoints/{id}/drain:
    post:
      description: Marks an endpoint as draining, preventing new requests from being
        routed to it
      parameters:
      - description: Endpoint ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Endpoint marked for draining
        "400":
          description: ID required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKeyAuth: []
      summary: Drain Endpoint
      tags:
      - endpoints
  /fingerprints:
    get:
      description: Returns all available fingerprint presets for TLS fingerprinting
      produces:
      - application/json
      responses:
        "200":
          description: List of presets
          schema:
            items:
              $ref: '#/definitions/github_com_kwilabs_straw-proxy-server_internal_server_dto.FingerprintResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKeyAuth: []
      summary: List Fingerprint Presets
      tags:
      - fingerprints
    post:
      consumes:
      - application/json
      description: Creates a new fingerprint preset or updates an existing one
      parameters:
      - description: Fingerprint preset
        in: body
        name: preset
        required: true
        schema:
          $ref: '#/definitions/github_com_kwilabs_straw-proxy-server_internal_server_dto.CreateFingerprintRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Created or updated preset
          schema:
            $ref: '#/definitions/github_com_kwilabs_straw-proxy-server_internal_server_dto.FingerprintResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKeyAuth: []
      summary: Create or Update Fingerprint Preset
      tags:
      - fingerprints
  /fingerprints/broadcast:
    post:
      description: Sends all fingerprint presets to all connected endpoints via message
        broker
      responses:
        "200":
          description: Presets broadcast successfully
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminKeyAuth: []
      summary: Broadcast Fingerprint Presets
      tags:
      - fingerprints
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer token authentication (format: "Bearer <token>")'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"